% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.ClientCompanyData}
\alias{analyze.ClientCompanyData}
\title{Analyze Portfolio Mix}
\usage{
analyze.ClientCompanyData(portfoliomix, startyear)
}
\arguments{
\item{portfoliomix}{Portfolio Mix analysis results prepared by \link{analyze.PortfolioMix} function}

\item{startyear}{Reference year of analysis}
}
\description{
Analyze Portfolio Mix
}
