% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.CorporateEconomy}
\alias{analyze.CorporateEconomy}
\title{Analyze corporate economy}
\usage{
analyze.CorporateEconomy(ald.master, iea.targets, startyear)
}
\arguments{
\item{ald.master}{Direct Owner Master dataset prepared by
\link{proc.prepareSectorMaster4analysis} function}

\item{iea.targets}{IEA Targets dataset prepared by
\link{proc.prepareIEATargets4analysis} function}

\item{startyear}{reference year for analysis}
}
\description{
Analyze corporate economy
}
