% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.MatchedClientsMixInclMaturity}
\alias{analyze.MatchedClientsMixInclMaturity}
\title{Analyze Matched Clients Mix including Maturity}
\usage{
analyze.MatchedClientsMixInclMaturity(loanbook, ald.master, startyear,
  years, all.lists)
}
\arguments{
\item{loanbook}{loanbook data with merged match results prepared for analysis}

\item{ald.master}{Asset Level Data Master dataset prepared by \link{proc.prepareSectorMaster4analysis} function}

\item{startyear}{Reference year of analysis}

\item{years}{number of consequent years (after startyear) to analyze}

\item{all.lists}{sets of various values for analysis purposes}
}
\description{
Analyze Matched Clients Mix including Maturity
}
