% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.meta.LoanbookMatches}
\alias{analyze.meta.LoanbookMatches}
\title{Meta-analysis on Loanbook matches}
\usage{
analyze.meta.LoanbookMatches(loanbook, ald.master)
}
\arguments{
\item{loanbook}{loanbook data with merged match results prepared for analysis}

\item{ald.master}{Direct Owner Master dataset prepared by \link{proc.prepareSectorMaster4analysis} function}
}
\description{
Meta-analysis on Loanbook matches
}
