% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.PortfolioCO2Intensity}
\alias{analyze.PortfolioCO2Intensity}
\title{Analyze CO2 Intensity}
\usage{
analyze.PortfolioCO2Intensity(portfoliomix, corp.econ.data, iea.targets,
  startyear, years)
}
\arguments{
\item{portfoliomix}{Portfolio Mix analysis results prepared by
\link{analyze.PortfolioMix} function}

\item{corp.econ.data}{Market Data derived from corporate economy based on
asset level data}

\item{iea.targets}{IEA Targets dataset prepared by
\link{proc.prepareIEATargets4analysis} function}

\item{startyear}{Reference year of analysis}

\item{years}{number of years to analyze}
}
\description{
Analyze CO2 Intensity
}
