% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.PortfolioMixTrajectory}
\alias{analyze.PortfolioMixTrajectory}
\title{Analyze Portfolio Mix Trajectory}
\usage{
analyze.PortfolioMixTrajectory(portfoliomix, iea.targets, startyear, years)
}
\arguments{
\item{portfoliomix}{Portfolio Mix analysis results prepared by \link{analyze.PortfolioMix} function}

\item{iea.targets}{IEA Targets dataset prepared by \link{proc.prepareIEATargets4analysis} function}

\item{startyear}{Reference year of analysis}

\item{years}{number of years to analyze}
}
\description{
Analyze Portfolio Mix Trajectory
}
