% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.TechnologyMix}
\alias{analyze.TechnologyMix}
\title{Analyze Technology Mix}
\usage{
analyze.TechnologyMix(portmixtrj, startyear, years)
}
\arguments{
\item{portmixtrj}{Portfolio Mix Trajectory prepared by \link{analyze.PortfolioMixTrajectory} function}

\item{startyear}{Reference year of analysis}

\item{years}{number of years to analyze}
}
\description{
Analyze Technology Mix
}
