% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{analyze.TPED}
\alias{analyze.TPED}
\title{Demand Analysis}
\usage{
analyze.TPED(loanbook, demand, startyear, years)
}
\arguments{
\item{loanbook}{loanbook data for analysis}

\item{demand}{total primary energy demand (TPED) proxy dataset}

\item{startyear}{Reference year of analysis}

\item{years}{number of consequent years (after startyear) to analyze}
}
\description{
Asessment across the Power and Fossil Fuels sector. Scenario
analysis (volume effect proxy) and Technology mix
}
