% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-datacheck-functions.R
\name{check.ds_cols}
\alias{check.ds_cols}
\title{Check completeness of dataset attributes}
\usage{
check.ds_cols(ds_name, ds_col_names)
}
\arguments{
\item{ds_name}{dataset name}

\item{ds_col_names}{vector of columns from dataset that needs to be checked for
completeness}
}
\description{
Procedure returns NULL if dataset complete or raises error if
  required columns are missing
}
