% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-matching-engine.R
\name{closestMatches}
\alias{closestMatches}
\title{Find cloasest matches}
\usage{
closestMatches(names, dictionary)
}
\arguments{
\item{names}{- vector of names to find in disctionary}

\item{dictionary}{- dictionary}
}
\value{
vector of indexes of dictionary entries
}
\description{
Function finds closest matches in a dictionary for the names provided
}
