% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-general-functions.R
\name{convert.naics2sc}
\alias{convert.naics2sc}
\title{convert NAICS -> Sector Classification}
\usage{
convert.naics2sc(naics, rules = pacta::data.naics2sc.rules)
}
\arguments{
\item{naics}{\itemize{
\item 6-digit NAICS Industry code
}}

\item{rules}{\itemize{
\item dataframe with conversion rules (example: \link{data.naics2sc.rules})
}}
}
\description{
Function takes list of NAICS codes and returns vector of
corresponding Sector Classifications. Additionally dataframe with conversion
rules in \link{data.naics2sc.rules} format can be provided.
}
