% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-5-analyze.R
\name{cross_join}
\alias{cross_join}
\title{Make cartesian product of two tables}
\usage{
cross_join(l, r)
}
\arguments{
\item{l, r}{- left and right hand side tables}
}
\description{
function returns combination of all records in \code{l} with all
  records from \code{r}. Columns from \code{r} follow columns from \code{l}
  in the result.
}
