% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.country.iso.codes}
\alias{data.country.iso.codes}
\title{Country names and ISO codes}
\format{An object of class \code{data.frame} with 286 rows and 2 columns.}
\usage{
data.country.iso.codes
}
\description{
Dataframe with country names and 3-character ISO codes.
}
\keyword{datasets}
