% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.dataset.req.attrs}
\alias{data.dataset.req.attrs}
\title{Required Dataset columns}
\format{\describe{
\item{Dataset}{Dataset name (as it is called in \link{data.dataset.req.files})}
\item{Attribute}{Exact process that requires data}
\item{Purpose}{Required dataset name}
}}
\usage{
data.dataset.req.attrs
}
\description{
Functions take data as an input and return data as an output. In order to be in control of data completeness attribute sets should be checked before processing starts. Completeness checks described in this data frame.
}
\keyword{datasets}
