% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.dataset.req.files}
\alias{data.dataset.req.files}
\title{Required data files}
\format{\describe{
  \item{Stage}{Processing stage}
  \item{Process}{Exact process that requires data}
  \item{Dataset}{Required dataset name}
  \item{Location}{Dataset location (where it is stored as a file). Some 
    parametrization is available to define DataInput, Data, Results folders 
    ({data-input}, {data-temp}, {data-result} respectively)}
  \item{File Name Pattern}{File name pattern}
  \item{Comment}{}
}}
\usage{
data.dataset.req.files
}
\description{
Table includes information about required data for different processing stages.
Data should be available before processing may be started.
}
\keyword{datasets}
