% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.index.regions}
\alias{data.index.regions}
\title{Index regions}
\format{a list of following pairs of vectors \describe{
\item{IndexUniverse, IndexUniverseColname}{Set of Indexes}
\item{Global, Global_ISO}{Countries of global region}
\item{MSCIWorld, MSCIWorld_ISO}{MSCI World}
\item{MSCIEmergingMarkets, MSCIEmergingMarkets_ISO}{MSCI - Emegring Markets}
\item{MSCIACWI, MSCIACWI_ISO}{}
\item{Stoxx600, Stoxx600_ISO}{}
\item{SP500, SP500_ISO}{}
\item{OutsideACWI, OutsideACWI_ISO}{}
}}
\usage{
data.index.regions
}
\description{
Different world regions described in terms of regions and countries.
Data contains lists of vectors, vectors represented in pairs of
Long Names (e.g. Country name) and ISO codes (Country ISO 2-char code)
}
\keyword{datasets}
