% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.name.reductions}
\alias{data.name.reductions}
\title{Name reduction rules}
\format{\describe{
  \item{From}{Original value to be replaced}
  \item{To}{Resulting value}
}
Values can be provided in any case, they are explicitely converted into lower
case (as a name) before use.}
\usage{
data.name.reductions
}
\description{
Name reduction (abbreviation) fules are used for name simplification. Data 
is represented in form of data.frame with two columns.
}
\keyword{datasets}
