% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.ownership.types}
\alias{data.ownership.types}
\title{Company ownership types}
\format{An object of class \code{character} of length 33.}
\usage{
data.ownership.types
}
\description{
Vector of company ownership types (like PLC, LLC, etc) is used to distinguish 
ownership type in company name during simplification.
}
\keyword{datasets}
