% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.sector.classification}
\alias{data.sector.classification}
\title{Sector Classification list}
\format{An object of class \code{character} of length 7.}
\usage{
data.sector.classification
}
\description{
Vector of Sector Classification Values
}
\keyword{datasets}
