% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.technology}
\alias{data.technology}
\title{Technology list}
\format{An object of class \code{character} of length 12.}
\usage{
data.technology
}
\description{
Vector of Technologies used in analysis
}
\keyword{datasets}
