% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-general-functions.R
\name{dup.test}
\alias{dup.test}
\title{Test vector for duplicates}
\usage{
dup.test(x)
}
\arguments{
\item{x}{\itemize{
\item input vector of values
}}
}
\value{
vector of logical values,
TRUE - element present more than once,
FALSE - element is unique
}
\description{
Function tests if vector of values and returns another vector of
logical values of the same length showing if element is unique
}
