% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-1-prepare-ald-for-matching.R
\name{empty.ALD4mathcing}
\alias{empty.ALD4mathcing}
\title{Empty asset-level data frame for mathcing}
\usage{
empty.ALD4mathcing()
}
\description{
Function defines data frame structure for asset-level data used in 
matching.
}
