% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-1-prepare-lbk-for-matching.R
\name{empty.LBK4mathcing}
\alias{empty.LBK4mathcing}
\title{Empty loanbook data frame for mathcing}
\usage{
empty.LBK4mathcing()
}
\description{
Function defines data frame structure for loanbook data used in
matching.
}
