% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-1-prepare-lbk-for-matching.R
\name{empty.ManualMatches4matching}
\alias{empty.ManualMatches4matching}
\title{Empty manual matches data frame}
\usage{
empty.ManualMatches4matching()
}
\description{
Function defines data frame structure for the manual matches data
added to the automatically generated matches
}
