% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{empty.SectorMaster}
\alias{empty.SectorMaster}
\title{Empty Sector master data frame}
\usage{
empty.SectorMaster()
}
\description{
Function defines data frame structure for Sector Master data used in 
analysis
}
