% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-datacheck-functions.R
\name{get.ds_file_count}
\alias{get.ds_file_count}
\title{Number of files for each dataset from requirements}
\usage{
get.ds_file_count(ds_name, files)
}
\arguments{
\item{ds_name}{Dataset name}

\item{files}{list of files to check if they serve as a source for dataset.}
}
\description{
Function checks vector of given files (with full path) and returns number of
files that provide data for specified dataset.
}
