% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-datacheck-functions.R
\name{get.ds_file_names}
\alias{get.ds_file_names}
\title{List of files for each dataset from requirements}
\usage{
get.ds_file_names(ds_name, files, substitutions)
}
\arguments{
\item{ds_name}{dataset name}

\item{files}{vector of file names}

\item{substitutions}{list of substitutions (list(keyword = value)), so
{keyword} entries can be replaced in vector of filenames}
}
\description{
List of files for each dataset from requirements
}
