% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-datacheck-functions.R
\name{get.ds_missing_cols}
\alias{get.ds_missing_cols}
\title{Find list of missing columns}
\usage{
get.ds_missing_cols(ds_name, ds_col_names)
}
\arguments{
\item{ds_name}{Dataset Name}

\item{ds_col_names}{vector of Dataset columns}
}
\description{
Function returns vector of missing columns (among given columns
  as an input parameter) in a given dataset. If all columns are present then
  function returns empty vector. Requirements are taken from packages internal dataset \link{data.dataset.req.attrs}
}
