% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.countFiles}
\alias{io.countFiles}
\title{Count files in folder}
\usage{
io.countFiles(path, pattern)
}
\arguments{
\item{path}{folder where to count files}

\item{pattern}{file-name pattern}
}
\description{
function counts files by given file-name pattern in specific
folder.
}
