% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.export.csv}
\alias{io.export.csv}
\title{Data export}
\usage{
io.export.csv(data, file)
}
\arguments{
\item{data}{data to be exported}

\item{file}{path to value; by default it is path to file (without extension)
where data is stored}
}
\description{
Function exports data into Key-Value store by given key. By
  default file system is used as KV store. Serialization format is CSV
  (american centric).
}
