% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.file.read}
\alias{io.file.read}
\title{read input and return to caller (generic method to read from KV store)}
\usage{
io.file.read(f, k)
}
\arguments{
\item{f}{function that reads value}

\item{k}{key that describes where value stored}
}
\description{
Function takes another function that performs read and location that tells
what exactly needs to be read.
}
