% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.file.write}
\alias{io.file.write}
\title{write output (generic method to write into KV store)}
\usage{
io.file.write(v, f, k)
}
\arguments{
\item{v}{value to be written}

\item{f}{function that writes value f(v, k)}

\item{k}{key that describes where value stored}
}
\description{
Procedure takes data (value), function that performs write operation and location that tells
what exactly needs to be written
}
