% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.getMissingFileAttr}
\alias{io.getMissingFileAttr}
\title{Get list of missing attributes of files}
\usage{
io.getMissingFileAttr(file, ds_name)
}
\arguments{
\item{file}{CSV file to check, contains full path to file (not only file name)}

\item{ds_name}{dataset name from \link{data.dataset.req.attrs} to retrieve
requirements}
}
\description{
Function returns vector of missing attributes in a file which 
  are necessary for processing
}
