% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.import.csv}
\alias{io.import.csv}
\title{Data import}
\usage{
io.import.csv(file, ...)
}
\arguments{
\item{file}{path to data; by default it is path to file (without extension)
where data is stored}

\item{...}{parameters of \code{read.csv} function}
}
\description{
Function imports data from Key-Value store by given key. By
default file system is used as KV store. Serialization format is CSV
(american centric).
}
