% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.load}
\alias{io.load}
\title{Load data object from dump}
\usage{
io.load(key)
}
\arguments{
\item{key}{location of value}
}
\description{
Fast using internal serialization format.
  Works best for temporary / intermediate data.
}
