% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.save}
\alias{io.save}
\title{Save data object from dump}
\usage{
io.save(value, key)
}
\arguments{
\item{value}{value to be stored}

\item{key}{location of value}
}
\description{
Fast save using internal serialization format.
Works best for temporary / intermediate data.
}
