% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-general-functions.R
\name{is_empty}
\alias{is_empty}
\title{Check if value is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{\itemize{
\item vector of values for check
}}
}
\value{
Logical value TRUE or FALSE; if input value is vector then function
returns vector of logical values
}
\description{
Function checks if provided value is NA or empty
}
