% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-general-functions.R
\name{is_null}
\alias{is_null}
\title{Replace empty value with alternative}
\usage{
is_null(x, y)
}
\arguments{
\item{x}{- vector of main values}

\item{y}{- vector of alternative values (if main value is empty)}
}
\value{
Either main value (if it is not NA) or alternative value
}
\description{
Function checks if provided value is NA and replaces it 
with alternative value
}
