% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-2-match-lbk-with-ald.R
\docType{data}
\name{match.score.perfect}
\alias{match.score.perfect}
\title{Perfect match score}
\format{An object of class \code{numeric} of length 1.}
\usage{
match.score.perfect
}
\description{
Perfect score - it is used for splitting match results
into perfect and thise which require verification
}
\keyword{datasets}
