% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-2-match-lbk-with-ald.R
\docType{data}
\name{match.score.threshold}
\alias{match.score.threshold}
\title{Match threshold}
\format{An object of class \code{numeric} of length 1.}
\usage{
match.score.threshold
}
\description{
threshold score - it is used for filtering out all match results
which are below thresold level.
}
\keyword{datasets}
