% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-matching-engine.R
\name{matchNamesByBlocks}
\alias{matchNamesByBlocks}
\title{Match names using blocking}
\usage{
matchNamesByBlocks(names, nblocks, dictionary, dblocks, threshold = 0.95,
  return = "closest")
}
\arguments{
\item{names}{\itemize{
\item name to compare against dictionary
}}

\item{nblocks}{\itemize{
\item vector that specifies blocks for names; should have the same
length as names
}}

\item{dictionary}{\itemize{
\item set of plausible values to chose from
}}

\item{dblocks}{\itemize{
\item vector that specifies blocks for dictionary elements; should
have the same length as names
}}

\item{threshold}{\itemize{
\item threshold value, minimum score to filter out results with
}}

\item{return}{\itemize{
\item what values to return from c("closest", "all"). Return type
"closest" works much faster and returns only 1 match per name with highest
score.
}}
}
\value{
dataframe of structure described as result of \link{matchNames}
function, extended with \code{Segment} column that represents block.
}
\description{
split names and dictionary values into blocks and make matching
within corresponding blocks
}
