% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta-package.R
\docType{package}
\name{pacta-package}
\alias{pacta}
\alias{pacta-package}
\title{pacta: Paris Agreement Capital Transition Assessment}
\description{
R package to perform Paris Agreement Capital Transition Assessment for corporate lending portfolio including matching and analysis module.
}
\author{
\strong{Maintainer}: Klaus Hogedorn \email{klaus@2degrees-investing.org}

}
\keyword{internal}
