% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-3-merge-matches-into-lbk.R
\name{pickBestMatches}
\alias{pickBestMatches}
\title{Pick best matches}
\usage{
pickBestMatches(matches, local.level, ald.levels, first.dupl = T)
}
\arguments{
\item{matches}{\itemize{
\item data frame with match results
}}

\item{local.level}{\itemize{
\item which local level of loanbook data is gouing to be merged
}}

\item{ald.levels}{\itemize{
\item vector of ALD data level in order of their priority
}}

\item{first.dupl}{\itemize{
\item how to treat first duplicate:
TRUE - keep it; FALSE - exclude duplicates completely
}}
}
\value{
dataframe with columns
"ID.local", "Level.global", "ID.global", "Original.Value.global"
}
\description{
Function pickes best matches from matches data frame provided.
"Best" is defined based on order of "Level.global" values provided in
\code{ald.levels} vector.
}
