% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe functions}
\arguments{
\item{lhs, rhs}{An R object and a function to apply to it}
}
\description{
Like dplyr, twodii4loans uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
