% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{proc.prepareAllLists4analysis}
\alias{proc.prepareAllLists4analysis}
\title{Prepare sets of values for Analysis purposes.}
\usage{

  proc.prepareAllLists4analysis(benchregions = pacta::data.benchmark.regions,
  indexregions = pacta::data.index.regions,
  countries = pacta::data.country.iso.codes,
  techlist = pacta::data.technology,
  sectorlist = pacta::data.sector.classification)
}
\arguments{
\item{benchregions}{list of benchmark regions (optional)}

\item{indexregions}{list of index regions (optional)}

\item{countries}{data frames with Country names and 2-character ISO codes
(optional)}

\item{techlist}{list of technologies to consider for analysis}

\item{sectorlist}{list of sectors to consider in analysis}
}
\description{
This function puts information about Regions and countries and some other
data (like technology lists) into one list of lists which then can be used in
analysis
}
