% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{proc.prepareIEATargets4analysis}
\alias{proc.prepareIEATargets4analysis}
\title{Prepare International Energy Agency data for analysis}
\usage{
proc.prepareIEATargets4analysis(scenarios, startyear, years)
}
\arguments{
\item{scenarios}{data frame with IEA Targets}

\item{startyear}{reference year for analysis}

\item{years}{number of consequent years (after startyear) to analyze}
}
\description{
Prepare International Energy Agency data for analysis
}
