% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{proc.prepareLoanbook4analysis}
\alias{proc.prepareLoanbook4analysis}
\title{Take matched loanbook and prepare dataset for analysis purposes}
\usage{
proc.prepareLoanbook4analysis(loanbook, exposure.type, customer.level,
  ald.levels, ald.ids, ald.names)
}
\arguments{
\item{loanbook}{matched loanbook data - contains pairs of attribues
with ALD Level, ALD ID, ALD Name}

\item{exposure.type}{defines what value to put as LoanSize (Outstanding,
Exposure)}

\item{customer.level}{Name of level on which mathcing is done}

\item{ald.levels, ald.ids, ald.names}{vectors of values that represent
reference to asset-level data. Vectors must be of same length as number of
rows in loanbook.}
}
\description{
Function adds columns necessary for analysis based on input
attribute values
}
