% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{proc.prepareSectorMaster}
\alias{proc.prepareSectorMaster}
\title{Prepare SectorsMaster dataset from specific sector datasets}
\usage{
proc.prepareSectorMaster(companymaster.ogc, assetmaster.ogc,
  companymaster.power, assetmaster.power, companymaster.auto,
  assetmaster.auto, companymaster.shp, assetmaster.shp, companymaster.cmt,
  assetmaster.cmt, companymaster.stl, assetmaster.stl, companymaster.avn,
  assetmaster.avn, startyear, years, techlist)
}
\arguments{
\item{companymaster.ogc, assetmaster.ogc}{Oil &  Gas data on company level and
asset level respectively}

\item{companymaster.power, assetmaster.power}{Electricity generation data on
company level and asset level respectively}

\item{companymaster.auto, assetmaster.auto}{Automotive data on company level
and asset level}

\item{companymaster.shp, assetmaster.shp}{Shipping data on company level
and asset level}

\item{companymaster.cmt, assetmaster.cmt}{Cement data on company level
and asset level}

\item{companymaster.stl, assetmaster.stl}{Steel data on company level
and asset level}

\item{companymaster.avn, assetmaster.avn}{Aviation data on company level
and asset level}

\item{startyear}{reference year for analysis}

\item{years}{number of consequent years (after startyear) to analyze}

\item{techlist}{list of technologies to include into result}
}
\description{
function takes datasets for each sector (2 per sector: on Asset and Company
levels) thansforms them into universal structure and combines information into one Sectors Master dataset
}
