% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-4-prepare-for-analysis.R
\name{proc.prepareSectorMaster4analysis}
\alias{proc.prepareSectorMaster4analysis}
\title{Prepare Asset Level Data for analysis}
\usage{
proc.prepareSectorMaster4analysis(sectormaster, all.lists, startyear,
  years, selregions)
}
\arguments{
\item{sectormaster}{All sectors data on company level and asset level
respectively (prepared by \link{proc.prepareSectorMaster})}

\item{all.lists}{sets of various values for analysis purposes}

\item{startyear}{reference year for analysis}

\item{years}{number of consequent years (after startyear) to analyze}

\item{selregions}{regions selected for analysis}
}
\description{
function takes datasets for each sector (2 per sector: on Asset and Company
levels) and combines information into one dataset suitable for analysis
purposes
}
