% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-2-match-lbk-with-ald.R
\name{process.flagNonPerfectMatches}
\alias{process.flagNonPerfectMatches}
\title{Extract non-perfect matches}
\usage{
process.flagNonPerfectMatches(names, sectors, scores)
}
\arguments{
\item{names}{vector of names (\code{Names} column in match results data frame)}

\item{sectors}{vector of sectors (\code{Segment} column)}

\item{scores}{vector of scores (\code{Score} column)}
}
\description{
function takes all matches produced by \link{process.matchLBKandALD} 
  function and flags only matches which require verification
}
