% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-2-match-lbk-with-ald.R
\name{process.flagPerfectMatches}
\alias{process.flagPerfectMatches}
\title{Flag perfect matches}
\usage{
process.flagPerfectMatches(scores)
}
\arguments{
\item{scores}{vector with match scores produced by \link{process.matchLBKandALD}
function (which contains \code{Score} column)}
}
\description{
Function takes match scores and returns logical vector of
perfect matches which can be used further without manual verification
}
