% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-0-generate-test-loanbook.R
\name{process.generateTestLoanbook}
\alias{process.generateTestLoanbook}
\title{Prepare test loanbook based on list of names, industry codes, countries}
\usage{
process.generateTestLoanbook(name, country, industry,
  industry_desc = NULL, rows = NULL, seed = NULL)
}
\arguments{
\item{name}{Company name}

\item{country}{Residense country}

\item{industry}{Company NAICS code}

\item{industry_desc}{Company NAICS description}

\item{rows}{Number of loans to Generate. In case number is provided companies
will be randomly picked up using \code{sample} function with
\code{replace = T} property.}

\item{seed}{Seed for rundom generation}
}
\value{
Loanbook data-frame with minimum required columnset
}
\description{
Prepare test loanbook based on list of names, industry codes, countries
}
