% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-2-match-lbk-with-ald.R
\name{process.matchLBKandALD}
\alias{process.matchLBKandALD}
\title{Match loanbook and asset-level data}
\usage{
process.matchLBKandALD(lbk4match, ald4match)
}
\arguments{
\item{lbk4match}{data frame prepared with \link{process.prepLBK4match}
function}

\item{ald4match}{data frame prepared with \link{process.prepALD4match}
function}
}
\value{
function returns bridge file produced by \link{matchNamesByBlocks}
function
}
\description{
function that matches loanbook data (prepared for matching)
and asset level data (prepared)
}
