% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-3-merge-matches-into-lbk.R
\name{process.mergeMatches2LBK}
\alias{process.mergeMatches2LBK}
\title{Merge matching results into Loanbook}
\usage{
process.mergeMatches2LBK(loanbook, matches)
}
\arguments{
\item{loanbook}{\itemize{
\item dataframe with loanbook data in stadard format
see \link{data.dataset.req.attrs} requirements in part of Loanbook
}}

\item{matches}{\itemize{
\item dataframe with match results
}}
}
\value{
Original loanbook dataframe plus 3 groups of columns related to
Asset, Borrower, Ultimate Parent level from ALD side.
Each group of columns consist of: ID, Level, Name. Columns of each group have
their own prefixes: Asset.link.to; DO.link.to; UP.link.to
}
\description{
Link each record of Loanbook to ALD data based on matches
produced by matching egine. Merging is done on several levels:
Assets, Borrowers, Ultimate Parents.
Loanbook assets can be matched to Assets, Owners, Parents (considering priority).
Loanbook loan-takers (Borrowers) can be matched to Owners or Parents.
Loanbook Ultimate parents can be matched to ALD parents only.
During data merge function follows priority of ALD levels:
\enumerate{
\item Assets, 2) Owners 3) Parents.
}
}
