% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-1-prepare-ald-for-matching.R
\name{process.prepALD4match}
\alias{process.prepALD4match}
\title{Prepare ALD data for matching}
\usage{
process.prepALD4match(ald)
}
\arguments{
\item{ald}{\itemize{
\item asset level data according to structure returned by
\link{empty.ALD4mathcing} function
}}
}
\value{
data.frame with column set:
Level, ID, OriginalValue, MatchValue, SectorClassification
}
\description{
function that prepares asset level data for matching. It
takes assets from ald database and produces dataframe of universal
structure for matching.
}
