% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-datacheck-functions.R
\name{susbtitute.string}
\alias{susbtitute.string}
\title{Substitute parts of string}
\usage{
susbtitute.string(str, subs)
}
\arguments{
\item{str}{string to modify}

\item{subs}{list of substitutions in format \code{list(name = value)}
where \code{name} is pattern and \code{value} is replacement.}
}
\description{
Substitute parts of string
}
