% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-1-prepare-ald-for-matching.R
\name{addALD4matching}
\alias{addALD4matching}
\title{Prepare dataset for matching}
\usage{
addALD4matching(levels, ids, values, sectors, simplify = FALSE)
}
\arguments{
\item{levels}{\itemize{
\item Level at which matching is performed (e.g. "Direct Owner", "Ultimate Economic Parent" etc.)
}}

\item{ids}{\itemize{
\item vector with cusomer ID
}}

\item{values}{\itemize{
\item vector with values to be matched
}}

\item{sectors}{\itemize{
\item vector with sector sclassification
}}

\item{simplify}{\itemize{
\item logical flag if names simplification should be applied (use only for names)
}}
}
\description{
Prepare dataset for matching
}
