% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta-package.R
\docType{package}
\name{pacta-package}
\alias{pacta}
\alias{pacta-package}
\title{pacta: Paris Agreement Capital Transition Assessment}
\description{
R package to perform Paris Agreement Capital Transition Assessment for corporate lending portfolio including matching and analysis module.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/pacta}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/pacta/issues}
}

}
\author{
\strong{Maintainer}: Klaus Hogedorn \email{klaus@2degrees-investing.org}

}
\keyword{internal}
