% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-matching-engine.R
\name{stringSim}
\alias{stringSim}
\title{String similarity}
\usage{
stringSim(string, dictionary)
}
\arguments{
\item{string}{\itemize{
\item value to compare against dictionary
}}

\item{dictionary}{\itemize{
\item set of plausible values to chose from
}}
}
\value{
Vector of numeric values between 0 and 1 where 0 - lowest similarity
and 1 - highest. Vector length is the same as dictionary length. Similarity
scores ordered in the same order as dictionary elements.
}
\description{
Function identifies similarity between provided string and
vector of strings called dictionary.
}
