% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-general-functions.R
\name{trim}
\alias{trim}
\title{Trim whitespaces}
\usage{
trim(x)
}
\arguments{
\item{x}{\itemize{
\item character string
}}
}
\value{
character trimmed string
}
\description{
Trim leading and trailing whitespaces from text value
}
